余分なファイルを削除

+++ setup.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 
-DOT_FILES=( .zshrc .zshrc.alias .zshenv .zlogin .gitconfig .gitignore .inputrc .tmux.conf .dir_colors .emacs.d)
+DOT_FILES=( .zshrc .zshrc.alias .zshenv .zlogin .gitconfig .gitignore .tmux.conf .dir_colors .emacs.d)
 
 for file in ${DOT_FILES[@]}
 do
-    ln -s $HOME/dotfiles/$file $HOME/$file
+    [ -s $HOME/$file ]  || ln -s $HOME/dotfiles/$file $HOME/$file
 done
